+++ src/assets/route-generate/routeGenerateUtils.ts
@@ -1,7 +1,7 @@
-import { DatabaseModel, FileToGenerate, GenerateProps, PromptsModels } from '@src/types/commands/route-create';
+import { DatabaseModel, FileToGenerate, GenerateProps, PromptsModels } from '@src/types/commands/route-generate';
 import { RepositorySkJson } from '@src/types/project';
 
-export abstract class RouteCreateUtils<V> {
+export abstract class RouteGenerateUtils<V> {
   abstract getFiles(props: GenerateProps): FileToGenerate[];
   abstract postGeneration(variables: V, props: GenerateProps): Promise<void>;
   abstract getVariables(repository: RepositorySkJson, route_path: string): V;